requirements.txt
streamlit
python-dotenv
httpx